refactor: reorganize project structure to remove TypeScript code
- Move Python code from iterm_mcp_python/ to root directory
- Update imports to work with the new directory structure
- Remove old TypeScript code and dependencies
- Update pyproject.toml to reflect new package structure
- Clean up project to focus only on Python implementation
- Maintain full functionality with the same API
