+++ __init__.py
@@ -1,0 +1,1 @@
+"""Test package for iTerm MCP."""
