--- iterm_mcp_python/__init__.py
@@ -1,7 +1,0 @@
-"""
-iTerm Model Context Protocol (MCP) integration.
-
-This package provides MCP tools for interacting with iTerm2 terminal sessions.
-"""
-
-__version__ = "0.1.0"
