--- package.json
@@ -1,37 +1,0 @@
-{
-  "name": "iterm-mcp",
-  "version": "1.2.3",
-  "description": "A Model Context Protocol server that provides access to the currently active tab of iTerm",
-  "homepage": "https://github.com/ferrislucas/iterm-mcp#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ferrislucas/iterm-mcp.git"
-  },
-  "author": "Ferris Lucas",
-  "bugs": {
-    "url": "https://github.com/ferrislucas/iterm-mcp/issues"
-  },
-  "type": "module",
-  "license": "MIT",
-  "bin": {
-    "iterm-mcp": "./build/index.js"
-  },
-  "files": [
-    "build"
-  ],
-  "scripts": {
-    "build": "tsc && node -e \"require('fs').chmodSync('build/index.js', '755')\"",
-    "prepublishOnly": "yarn run build",
-    "watch": "tsc --watch",
-    "inspector": "npx @modelcontextprotocol/inspector build/index.js",
-    "debug": "ts-node --esm test/CommandExecutor.test.ts"
-  },
-  "dependencies": {
-    "@modelcontextprotocol/sdk": "0.6.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20.11.24",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.3.3"
-  }
-}
